module/cyanide: players import ok
Update issue 182
It's now possible to import a team and its players. A coach can import a self owned team and a Commisioner can import a team for anyone.
If the team already exist it is updated.
Todo:
 * update team goods
 * update player statistic
 * update player properties
Open point: I think it could be better that module hacks in main files (sections.php, admin.php...) could be manage using one entry point. For example the section import could call a function 'get_module_import' in charge to retrieve all import capabilities from module. The idea is to avoid to modify main files for modules needs.